[release-4.4] e2e: add extended tests for topology manager
The tests will skip if topology manager is not enabled
on the cluster, with single-numa node policy.
If it is enabled, will check the pod resources are usable
and properly aligned on NUMA node.

This is a superset of the e2e tests found on kubernetes.
A key difference is that these tests run against a full cluster,
while the e2e node tests runs in a reduced environment.

release-4.4 backport note
- ported back to the 1.17 client API (no `ontext` in calls)

Signed-off-by: Francesco Romani <fromani@redhat.com>