router/metrics: Fix haproxy_server_max_sessions test
This assertion currently flakes a lot. (I believe) the flake occurs
when the router reloads and the GET for the latest metrics reads
zero's because there has been no updates.

It may be better to mark this test as serial. This commit retries the
assertion as opposed to dealing with it as a one-shot.

The expectation is that haxproxy_server_max_sessions needs to be
greater than zero over two successive scrapes once the first scrape
shows an increase. However, this assertion can still fail depending on
when the metrics are refreshed and when the router reloads, which is
every 5s. In my testing I have seen this to be more resilient but
there are periods when running the full openshift/conformance tests
that this continues to fail.