tests: extended: fix TM regression test
The existing regression test which wants to cover against regressions
on RHBZ#1813567 is actually doing the reverse check than it should.
We noticed after having the test run for few days on environment kindly
provided by nfvpe team.
The likely cause of why this mistake sneaked in is that I wrongly
configured the original environment on which I tried the test.

This patch fixes the issue and puts extra care in documenting and
checking the intent of the original bugzilla.

Signed-off-by: Francesco Romani <fromani@redhat.com>