pkg/test/ginkgo: Add 'weight' property to failure JUnit <testcase>
This lays the groundwork for having the job-detail page rank
failing-case severity.  "failure" cases are the most important, and
users who are frustrated by their tests not passing should focus on
resolving those issues.  "flake" cases did not sink this particular
run, but should be addressed in the absence of "failure" cases to help
drive down sporadic stability issues.  "informer" cases do not need to
be "fixed", but might provide some context that helps understand
cluster health or root causes of failure and flake cases.

The upstream test-case structure already has a 'properties' propery
[1] and has since they started using JUnit [2].  Docs in [3] for which
I have a fixup in flight [4].

[1]: https://github.com/kubernetes/test-infra/blob/2eab27db6e653d809cd48ad06a35462f1aa3cf0e/gopherage/pkg/cov/junit/xmlwriter.go#L47
[2]: https://github.com/kubernetes/test-infra/pull/9887/commits/ce50d4c65c957286756b26f8baf3c786270b8077#diff-33ff1c70f03ec46f8567ad05cacf5254R47
[3]: https://github.com/kubernetes/test-infra/blob/2eab27db6e653d809cd48ad06a35462f1aa3cf0e/gubernator/README.md#test-properties-optional
[4]: https://github.com/kubernetes/test-infra/pull/17794