test: fix network-plugin-related skips
The old code allowed doing both "[Skipped:Network/OpenShiftSDN]"
("Skip on OpenShiftSDN") and
"[Skipped:Network/OpenShiftSDN/Multitenant]" ("Skip on OpenShiftSDN in
Multitenant mode"), and the test annotations made use of this. But the
current code lost that distinction and ended up skipping the "only
skip in Multitenant mode" tests even in NetworkPolicy mode (meaning we
were always skipping the NetworkPolicy tests on openshift-sdn).