WIP: test/e2e/upgrade: Split update testing into per-hop cases
"Cluster should remain functional during upgrade" is a big bucket, and
it's hard to figure out what's going on when it fails based on just
the test-case [1].  This commit pulls the setup steps out into their
own case (which is unlikely to fail, but still).  And it puts each hop
in its own case, including the source and target version.

FIXME: break out major/minor so we can aggregate in Sippy, etc.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1843183