UPSTREAM: <drop>: preserve old health/metrics behavior
Tweak the upstream commits to preserve the old behavior for kube-proxy
health and metrics. In 4.4 we will get the upstream fix for health,
and will need to ensure that our monitoring is fixed to use different
metrics.

MANUAL EDIT:
the metric metrics.SyncProxyRulesLastTimestamp does not exist in the version of kube in 3.11 so I need
to remove it

change SyncProxyRules to forceSyncProxyRules() in proxier_test.go