router/grpc-interop: switch to WaitForPodRunningSlow
The builder pod is taking a long time to complete in CI. In local
testing (i.e., on a quiescent cluster) it takes between 45s-60s to
complete. In CI this is taking much longer so switching to
WaitForPodRunningSlow which has a longer timeout (currently 15m).

To help reduce flakes:

- Increase internal test timeout to 5m
- Reduce number of test iterations from 5 to 2.

We want to run more than once but 5 is probably unnecessary in CI so
now reducing to just 2 iterations.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1845434