Measure container restarts over an interval instead of absolute
We run these tests after disruptive events where restarts are
necessary. Instead of looking at absolute restart count, capture
the number of initial restarts and then compare subsequent runs
to ensure the count is smaller. We double the time interval and
halve the number of failing restarts in order to ensure we capture
slow restart failures (full backoff is 5m, but we can catch slow
but not limited backoff).