test/extended/prometheus: During transition, allow HTTP(S) for CVO
Currently the master CVO serves only HTTP.  There is a pull request in
flight (linked from the Go comment) to declare an HTTPS target in the
ServiceMonitor.  Allowing both, as I'm doing in this commit, will
allow us to land this test-suite change now.  Then we can land the CVO
pull request.  And then we can drop the '?' to require HTTPS going
forward.