Bug: Cockroach  2.1.11 operator was removed
This commit updates an olm e2e test that ensures operator subscription
works. The test currently relies on an external repository of operator
metadata to generate that subscription. An update removed the 2.1.11
version from the current update graph, so this test is broken. As a
temporary solution, updating the test to point to the latest available
version. In the future, this test should either ship its own manifests
(so that it is not dependent on an external content souce) or be
removed entirely.