test: Report disruption flakes correctly in junit
The Flakef function currently results in upgrade subtests being
flagged as full failures, when they should be logged as "passed,
but flaky in some form". If we pass but capture a flake, output
two TestCase entries into the JUnit XML (one passing, one failing)
which correctly triggers downstream tools to show the result as
a flake, not a failure.

Should dramatically simplify viewing of Upgrade jobs.