extends StartOpenShiftAPIMonitoring to monitor GET and PATCH requests
It seems that retries are a popular way of increasing high availability of the system in the face of transient and random errors.
However, knowing which requests are worth retiring is hard.
Some requests are simply not safe because they might have side effects, others might be retried by different components (client vs server).

In order to understand the impact on availability under a disruption, we decided to monitor PATCH requests to the server to see if/how they differ from GET requests.