patch net/http2 library to fix the graceful shutdown of HTTP2 connections
this patch fixes net/http2 library to properly manage the connection state,
this allow to proper shutdown sequence to occur.
as a result Server.Shutdown will wait for active connection to finish before shutting down the HTTP server

4d8de76-1.13.patch was prepared based on https://go-review.googlesource.com/c/go/+/240278
the original patch targets the master branch and didn't apply to golang 1.13.