test: Report an error message when a pod fails in a cmd test
The logic for testing for pod failure lost the reason why, so no
debug output was possible. Instead, update the one shot pod to
capture fallback logs and then have the executor report that as
the error message.

E.g.

    Jul 22 20:13:43.152: INFO: pod "test-cmd" errored trying to run the command: <nil>

because err was nil in the context it was invoked