Add a new FIPS test
See discussion in
https://github.com/openshift/release/pull/10488
and https://bugzilla.redhat.com/show_bug.cgi?id=1861095

This test replaces the bash code in the release repo
with a more proper test here.

While here I noticed that the topology tests had some code
that reused the MCD as a handy privileged pod; extract
that to the toplevel utils and use both here and there.