Add --node-name and --node-ip flags to sdn
Currently openshift sdn tries to figure out the node name from the
hostname, and the IP address through a DNS query of that hostname.

With this code we're allowing to pass a flag to override both values.

After this is done we'll modify openshift-ansible to fill these flags
with environment variables that read fieldRefs.