Fix incorrect retrieval of sandbox pod
PR: https://github.com/openshift/origin/pull/23522 fixed syncing
initial SDN state for hostport pods. There was however a mistake in
that patch, namely: it overlooked that existingSandboxPods was a map
of sandboxID -> PodSandbox and was retrieving the Sandbox using the
"podKey" (namespace/name). This lead to other initialization errors
when the SDN restarts. This patch reverses the logic and fixes the
issue.

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>