Detect and report unexpected leader elections
Under ideal conditions, the following etcd term/election timeline is expected:

term  elections  reason
1     -          no bootstrap member
2     1          bootstrap member elected
3     2          bootstrap member removed, master elected
4     3          new etcd pod rev to drop bootstrap member from config

Elections 1 and 2 occur during etcd pivot and before Prometheus is scraping any
metrics, and so will be invisible. Election 3 is the first election that should
be collected in the metrics data. Any other elections are suspicious and could
indicate a problem (e.g. IO contention, packet loss) that we want to
investigate.

So, only 1 leader change is expected to be observed unless the test is either
disruptive or an upgrade.

This change adds a new monitor which adds a synthetic flake when unexpected
leader elections are detected so that we can identify and analyze CI runs in an
effort to reduce or eliminate such elections.

If in the future global CI Prometheus metrics are aggregated and made available
for analysis, this monitor can probably be removed.