test: Allow ginkgo tests to declare a flake
Declare that a passing run actually failed by having a standard
e2e test invoke the result.Flakef function. The test will continue
and be treated as a pass, but the JUnit generated for the suite
will continue a failure run (with the message from Flakef) and a
passing success run.