Remove broken machine-api taint test
This test suite is currently checking for labels
and taints on nodes created by the machine-api.
Other components are adding taints to existing workers
which are sometimes selected to scale down during
testing.  This test does not allow for any modification
of labels and taints.

The machine-api is not the sole or final manager of
labels and taints.  We have coverage in our own
test suite that the machine-api adds labels
and taints appropriately, this test is unnecessary
here and does not add cross-component signaling.