images/router: Add ART specific Dockerfile
PR #25542 introduced a change that allows us to pickup a specific
version of HAProxy during a CI build. Unfortunately this broke ART
builds as there is no external network access that would allow access
to the RPM files.

This commit reflects the state of the Dockerfile as it was before
25542 merged.

There is a complementary commit for ocp-build-data:
  https://github.com/openshift/ocp-build-data/pull/736

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1894011