e2e: node: explicit image assignment for tests
Pin the busybox image setting a known-good tag.
It was not done previously because of an oversight, no other reason.
Generalize the code which creates the testing pod to make the
code more flexible and a bit easier to read.

Any public-available, reasonably recent (>= 1.21) busybox image is fine.
We use the same image other extended tests (DNS) are already using.
This solves the issues for all but one tests: the connectivity test.

One of the e2e topology manager tests wants to run a basic network
check between two NUMA-aligned pods, to get a basic signal if
the network is working. This is done to test that resource alignment
is not preventing the network to work.

This test is intentionally minimal: proper network check is demanded
to other testsuites (e.g. the SRIOV testsuite), but we need some basic
coverage in this testsuite and we have this intentional, minimal,
overlap.

To do the minimal testing, the test what to ping two resource-aligned
pod from each other. To do so we need a container image which contains
the ping tool and which not require any special privilege.

So, this patch requires the pull URL of the network-check image
to be supplied using a (documented) environment variable. This way:
1. users can use existing images not part of openshift
   (quay.io/openshift-kni/cnf-tests) which they may want to use to
   validate their cluster anyway
2. users can easily opt out from this test (do not supply the env
   variable to skip the test)
3. in the future we can provide a very simple base image in from
   openshift itself to fill this role.

Bug-Url: https://bugzilla.redhat.com/1887488
Signed-off-by: Francesco Romani <fromani@redhat.com>