test/extended/operators: Require CPU and memory requests
The outgoing QoS check landed in aeb8caad73 (tests: extended: add pod
QoS validation check, 2019-05-06, #22787).  But while setting a
request for a single type (e.g. just CPU) on a single container is
enough to move a pod from the BestEffort class to the Burstable class,
it's not enough to protect a pod from eviction if a separate,
non-requested resource is constrained on the node.  Because all
containers will consume both memory and CPU, require, as part of
OCP-core policy, all OCP-core containers to set requests for both
resource types.  Future work can compare the requested values with
measured consumption, but I'm punting on that for now.

Generated by manually editing gos.go and running:

  $ hack/update-generated.sh