Bug 1895093: Verify builds can mount proxy trustedCA
Verify builds can mount the cluster's custom PKI trust bundle if one is
set on the cluster proxy configuration. When `mountProxyTrustedCA` is
set to `true`, builds should add the container's trust bundle as a
read-only mount. The custom PKI should be readable during the build
execution, but should not be present in the resulting image.