Re-enable test/extended/images/signatures.go
Don't read the image to be signed from docker.io.
Instead, sign the just-built signer image (which is unique, so
there always should be enough free signature slots left).

Send the required credentials to the source registry.

Also use the injected service CA instead of disabling TLS.  I don't know
whether it is supposed to work like that (per
https://github.com/openshift/openshift-docs/blob/enterprise-4.1/release_notes/ocp-4-1-release-notes.adoc#service-ca-bundle-changes
the path is deprecated) but the same path is already assumed to exist
by the preceding (oc login).

Signed-off-by: Miloslav Trmač <mitr@redhat.com>