test: Prometheus helper should use thanos-querier
This ensures that default queries aggregate results from the running
instances (during upgrades) and helps us test invariants like "there
is always an alert firing even during upgrades". Note that this makes
reconstructing a test more complex because we can no longer exactly
reproduce a query (by downloading the prometheus data from a job) but
that was already true because we would get round robined to one of the
instances and some data is implicitly missing.