test: Don't allow alerts to fire during a test run
The alert test has not been working and a PR was able to merge that
was firing alerts during the test run.

Fix the test logic based on:

1. Remove the filter on KubeAPILatencyHigh - no longer needed in 4.6+
2. Remove the filter on KubePodCrashLooping - no longer needed in 4.6+
3. Sort the metric by number of seconds the alert was firing

The bug that allowed the merge was the filter for KubePodCrashLooping,
to omit a series we must use `unless X`. If we reintroduce a filter we
should use that syntax.