test: Fail if a daemonset has a different upgrade strategy than default
Now that we have fixed all issues related to network upgrades by
moving OVS to the node, we can bulk upgrade daemonsets on the nodes.

Daemonsets shipped with the platform must be able to upgrade without
disruption to workloads. Daemonsets that are in the data path must
gracefully shutdown and redirect workload traffic, or mitigate outage
by holding requests briefly (very briefly!). Daemonsets that are part
of control plane (anything that changes state) must already be able
to be upgraded without disruption by having any components that call
them retry when unavailable up to a specific duration. Therefore all
workloads must individually be graceful, and so a core daemonset can
upgrade up to 10% of pods at a time. In the future, we may allow this
percentage to be tuned via a global config, and this test would
enforce that. For daemonsets that cannot disrupt the workload, they
may choose to use maxUnavailable 33%, which is equivalent to our
zone failure detection behavior.