Skip tests that rely on Machine API when Machine API is not installed
We have a number of tests that use the Machine API, but also topologies 
of deploying OpenShift that do not install the Machine API (eg IBM). To 
be able to support this going forward, ensure that all tests check for 
the presence of Machine API before starting the test. If the Machine API 
namespace doesn't exist, skip the test.