test: Graceful termination invariant should separate pre-suite events
Termination of the kube-apiserver outside of the suite that is currently
running should be a flake, and when the test infra is told to ignore
events before a specific time it should do so. Add the test into
the upgrade invariants since this would exclude the test from checking
upgrades otherwise.