DO NOT MERGE: Test whether multi-az test is impacted by deleting pods
Working hypothesis is that deleting pods are being taken into account
by the createBalancedPodForNodes method and then the nodes are left
slightly unbalanced, which results in workloads being steered to the
less utilized nodes. Observed 6 pods on each node that were from
previous tests, so there's a good chance that this reduces the potential
for churn.

Will test with this repeatedly.