test/e2e/upgrade: Relax 'too long' soft timeout for rollback jobs
durationToSoftFailure was added in 4447a19e86 (allow longer upgrade
times to run tests, but continue to fail at 75 minutes, 2020-08-12, #25411),
but didn't get the 2x on rollbacks we'e been adding to maximumDuration
since a53efd5e27 (Support --options on upgrade tests to abort in
progress, 2019-04-29, #22726).  That's recently been causing the
cluster-version operator's A->B->A rollback CI jobs to time out [1].
This commit catches durationToSoftFailure up with the "2x on
rollbacks" approach, and also mentions "aborted" in messages for those
types of tests, to help remind folks what's going on.

An alternative approach would be to teach clusterUpgrade to treat
rollbacks as two separate hops (one for A->B, and another for B->A).
But that would be a more involved restructuring, and since we already
had the 2x maximumDuration precedent in place, I haven't gone in that
direction.

[1]: https://github.com/openshift/cluster-version-operator/pull/514#issuecomment-801242495