pkg/synthetictests/operators: Include the reason string in testOperatorState
Extending the the logic from 6ec446bad7 (flake on operator state
changes during e2e, 2021-02-23, #25918) to include the reason, since
that slug both makes it easier to find the relevant operator code and
makes it easier to match CI flakes with seen-in-the-wild Telemetry
(where we get the status and reason, but not the message).

I'm also moving some string types into structured config/v1 types.
Narrowly, this makes it easier for GetOperatorConditionStatus to
return all its metadata without a big-list-of-strings API.  More
broadly, this drops things like "assume status is a boolean, and
neglect Unknown and unrecognized values".  This also means that we
can't use !status to guess at what the previous status was, so I've
dropped that and now only talk about the data we actually have.