[release-4.5] Bug 1939550: Update tests to use Ruby 2.6
Update orign tests to use Ruby 2.6 instead of Ruby 2.5.
github.com/openshift/ruby-hello-world was updated to use Ruby 2.7,
which breaks many tests that were based on Ruby 2.5.

Tests which relied on the Dockerfile in ruby-hello-world are disabled
because it explicitly uses the ruby 2.7 imagestream, which is not
available on OCP 4.5.

Other changes:

- Fix check on build args warning
- Test that checks if we infer the origin imagestream tag was switched
from a centos image to the ruby imagestream.