test: add baremetal provider
Baremetal IPI tests currently run without any provider set. While we're not so
interested in the cloud specific parts, we at least need a null provider
because loading the config also loads which SDN framework we're using,
and allows us to skip specific tests when using OVN.  Currently the
tests that should be skipped on OVN, aren't -- they're filtered manually
by passing a list of tests to skip.