test: remove apk add tcpdump to internal ports test
The connectivity test for ports 9000-9999 tries to install tcpdump via
apk, and this appears to only be used for debugging purposes. In order
to run these tests disconnected, we can't have dependencies running
in the container that try to reach out to the internet. Currently this
test fails on IPv6 disconnected installs.