test/extended/router: skip http2, grpc-interop and http2 on AWS
AWS/UPI jobs are failing on http2, grpc-interop and h2spec tests
because deleting the ingresscontroller which was explicitly created
for the test takes more than 10 minutes[1] and that (plus the overall
test time) exceeds the amount of time the test is allowed to run for
before being marked as failing.

This commit skips tests on AWS until we fix BZ #1912413.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1912413.