extends monitoring of Kube API servers by hitting NodeIPs directly
Already existing tests send traffic to the external load balancer.
The issue we see is when the tests fail we cannot tell if it was a misconfigured load balancer or a broken API server.

This PR reuses already existing tests both for new and reused connections by swapping the host address in the rest configuration for all kubernetes endpoints.