test: Add the Cinder client to the test image
Multiple tests require the Cinder client to be installed when run
against OpenStack.

In the step-registry CI structure, the "test" ref is shared across
platforms; it is therefore not desirable to run the OpenStack tests
using a separate image.

This change adds the Cinder client to the container image used to run
the tests in the CI.

Implements OSASINFRA-2308

cf. "[sig-storage] In-tree Volumes [Driver: cinder] [Testpattern:
Inline-volume (default fs)] volumes should store data
[Suite:openshift/conformance/parallel] [Suite:k8s]"