test: Record disruption events in a way that shows up in the log
When a disruption test records disruption, those should be output
to a file that can be read at the end of a run and merged into the
final output. This allows upgrade test disruption data to show up
in the intervals deck.

Stop using ARTIFACT_DIR directly and use the provided junit dir
(which should control all output from an openshift-test run).