Bug 1961367: Add Cinder client to the test image
Multiple tests require the Cinder client to be installed when run
against OpenStack.

In the step-registry CI structure, the "test" ref is shared across
platforms; it is therefore not desirable to run the OpenStack tests
using a separate image.

This change adds the Cinder client to the container image used to run
the tests in the CI.

Implements OSASINFRA-2308

cf. "[sig-storage] In-tree Volumes [Driver: cinder] [Testpattern:
Inline-volume (default fs)] volumes should store data
[Suite:openshift/conformance/parallel] [Suite:k8s]"

This is a cherry-pick of 559db12abbe938db5d726aa196901e7d5ef47c12
(https://github.com/openshift/origin/pull/26131) modified with the fix
for the architectures where the openstack repository is not available
5c861b50e95153cf2c56a34f1cf29790c25eb10c
(https://github.com/openshift/origin/pull/26170)