Flake when SCCs are not associated yet
https://bugzilla.redhat.com/show_bug.cgi?id=1961204 was punted
to 4.9 release, however occasionally we see multus not
getting SCCs in time causing auth issues. This commit
causes the test to flake in such scenarios instead of
failing. This is a temporary measure to get better
signal for 4.8 upgrades. We need to remove this
when the associated bug has been fixed