test/extended: Ensure all containers in the OLM deployment resources contain the IfNotPresent image pull policy
Update the test/extended/operators/olm.go and ensure that all containers
in the OLM deployment resources contain the IfNotPresent image pull
policy.

Previously, this test would make the assumption that all the OLM-related
deployments contain a single container, which won't always be the case
in the future. This updates that test to split the output from the oc
command into array elements, checking each element to ensure it matches
the correct image pull policy.