Skip more NetworkPolicy tests
The new "Netpol" suite attempts to retry if a test fails the first
time because it started probing before the plugin fully implemented
the policy. But the retry code is broken and the tests will still fail
if they failed the first time, even if they succeed the second time.