Bug 1977454: Use nodejs to test service connection
Redis has tighter constraints on client HTTP calls to prevent malicious
attacks. Only official redis clients are guaranteed to receive reliable
HTTP responses. The build test for in-cluster service connections does
not use a redis client, resulting in a test HTTP curl request receiving an
empty repsonse.

This updates the in-cluster service test to use a "hello world" nodejs
application and service. The use of the upstream Redis image is also
removed.