disruptive: output millis when disruption is < 1s
Currently when there's a disruption that's less than a second, we
truncate it to `0s` resulting in odd output like a failure with the
message `Frontends were unreachable during disruption for at least 0s of
1h17m57s`.

This change leaves the duration unchanged if less than a second, so
output will now say `during disruption for at least 382ms of 1h17m57s`.
For values > 1s, the current behavior will stay by rounding down.