Fix a lock contention problem in the proxy
Using the mutex in proxier.Sync() to protect proxier.changesPending
means that a call to proxier.Sync() can end up blocking waiting for a
call to proxier.syncProxyRules() to finish, which can take a long
time.

In particular, in a large cluster Sync() will get called thousands of
times during startup, and some of those calls will block waiting on
syncProxyRules() calls, which may take a few seconds, and "thousands
of times" times "a few seconds" can add up to 10 or 20 minutes of
delays in synchronizing.

Fix this by using atomics to set/get changesPending rather than
fighting with syncProxyRules() for the mutex.