Add generic internet egress check alongside API endpoints
There's a possibility the API disruption tests are failing because of
the environment inside the cluster from where openshift-tests is running
instead of on the target cluster under test.  We've noticed an uptick in
failures across multiple releases starting around the same time, on
separate platforms, with no obvious commits to point to.

This adds a check for a generic internet URL (neverssl.com is small, and
fit-for-purpose as a general check for connectivity).