CustomNoUpgrade should not fire no-upgrade failures
CustomNoUpgrade should be treated just as TechPreviewNoUpgrade, and
should not be a condition that triggers no-upgrade failures in tests.
This is useful in cases where we want to test a specific set of feature
gates in isolation from other tech preview feature gates.

Specifically we are using CustomNoUpgrade to test
external-cloud-provider CCM promotions in isolation in
https://github.com/openshift/release/blob/992f297ef8a1ba749f3a0b9b40438696f18a1bc4/ci-operator/step-registry/ccm/conf/pre-create-feature-gate/ccm-conf-pre-create-feature-gate-commands.sh#L22

Co-Authored-By: Matthew Booth <mbooth@redhat.com>