Increase threshold for scc failures flake
We're seeing increased rates of failure for pod creation failures, and
the BZ is currently being investigated. This changes the test to flake
instead of fail on the 80th percentile, which seemed like as good a
choice as any. Beyond that, the failures look more catastrophic with
more than just this test failing.

```
$ curl https://search.ci.openshift.org/\?search\=pods+failed+before+test+on+SCC+errors\&maxAge\=168h\&context\=1\&type\=junit\&name\=4.9\&excludeName\=\&maxMatches\=5\&maxBytes\=20971520\&groupBy\=job 2>/dev/null \
| egrep -oh "([0-9]+) pods failed before test" | \
cut -f1 -d" " | Rscript -e 'summary (as.numeric (readLines ("stdin")))'

   Min. 1st Qu.  Median    Mean 3rd Qu.    Max.
  3.000   4.000   8.000   8.558  11.000  28.000

$ curl https://search.ci.openshift.org/\?search\=pods+failed+before+test+on+SCC+errors\&maxAge\=168h\&context\=1\&type\=junit\&name\=4.9\&excludeName\=\&maxMatches\=5\&maxBytes\=20971520\&groupBy\=job 2>/dev/null \
| egrep -oh "([0-9]+) pods failed before test" \
| cut -f1 -d" " | Rscript -e 'quantile (as.numeric (readLines ("stdin")), probs=c(0.80))'
14
```