test/extended/router: Enable edge-terminated gRPC
Enable the grpc-interop with edge-terminated route test, using the
service's spec.ports[*].appProtocol field to specify that the backend pod
expects h2c (cleartext HTTP/2) connections.  Add a test that uses h2c on
both the frontend and the backend.

This commit is related to bug 1826225.

https://bugzilla.redhat.com/show_bug.cgi?id=1826225

* test/extended/router/grpc-interop.go: Add a test that uses h2c on the
frontend and backend.  Delete the skip for the edge-terminated route.
* test/extended/testdata/router/router-grpc-interop-routes.yaml: Add a
non-TLS route.  Use target port 1110 for the edge-terminated route.
* test/extended/testdata/router/router-grpc-interop.yaml: Add port 1110 for
gRPC over h2c to the grpc-interop service and pod container spec.
* test/extended/testdata/bindata.go: Update.