images: port image signature workflow test to OCP4/UBI8
This fixes portability issues with this test, given that
origin-control-plane:v3.11 (besides being obsolete) is only available
for x86_64.  Also, copying the oc binary from the machine running the
test into a container on the cluster would cause the test to fail when
run on a different architecture, or when oc was built for a newer glibc.
There was also no guarantee that the oc found on the test machine
matched that of the payload being tested.

Instead, this uses the payload cli container, which is based on RHEL 8.
Because 4.y CI builds do not have any publicly accessible yum repos, UBI 8
is added manually to get skopeo and its dependencies. The rest of the
changes are a consequence of adapting to the newer versions of gpg2 and
skopeo.