Bug 2008114: Add role label to jenkins imagestream
Add the role label to the Jenkins agent imagestream so the sync plugin
can pick up changes. Backport of changes for BZ 1925524.

Note that this change includes namings that violate inclusive language
standards. This is to maintain backwards compatibility with prior
versions of the sync plugin. Future versions of the sync plugin will
support more inclusive namings.