bug 2011924: Fix inaccurate test error for NodeReady multiple times.
In some cases (particularly on Azure), a node may take longer to come up
initially and actually emit it's first NodeReady event after the CI has
started the upgrade. When this happens the test would record that first
NodeReady, then the second NodeReady after the node was upgraded, and
throw an error the test itself said should be impossible.

To work past this we will disregard a NodeReady prior to seeing a
NodeNotReady, allowing us to more safely ensure we're not seeing a node
flap *after* it's attempted it's upgrade.